Prevent exiting content from reappearing on slow page loads
When the response is slow, the animation class name was being removed
from the element, but it wasn't being replaced by the new body in time.

As a result, the departing content would reappear. The solution is to
avoid removing the animation class name until after the page has loaded.
It's also important that the `animation-fill-mode` is set to `forwards`
so that the styles defined by the last keyframe remain applied to the
element.