Wait for enter animations in beforeRender
Scenario:
1. Click on "Very Slow".
2. Click "Back".
3. Click "Very Slow" again.

A preview will be rendered immediately, but it's entrance won't be
ainmated. When the response finally comes back from the server, we'll
see an animated entrance of that content.

This was happening because animations were rendering on load instead of
render. When a preview exists in the cache, `before-render` and `render`
are called twice. We need to run the entrance animation in render, but
we need to pause the rendering of the server response until that
entrance animation is complete.